+++ CHANGES.markdown
@@ -1,5 +1,10 @@
 ## Changes in next
  - Drop GHC 6.12 (and `base-4.2.0.0`) compatibility
  - Fix Windows, GHCJS build
- - Add `Typeable` instances for `CodePageArrays`, `CompactArray`, and
-   `ConvArray` on Windows
+ - `Read` instance for `Fixed`
+ - `Eq` instances for `Chan`, `QSem`, and `QSemN`
+ - `Applicative` instances for strict and lazy `ST`
+ - `Typeable` instance for `SampleVar`
+ - `Applicative` and `Alternative` instances for `ReadP` and `ReadPrec`
+ - `Typeable` instance for `KProxy`
+ - `Typeable` instances for more data types in `GHC.`-prefixed modules